present terms/privacy consistently
