+++ Kiosk/App/AppDelegate.swift
@@ -7,9 +7,10 @@ let log = XCGLogger.defaultInstance()
 public class AppDelegate: UIResponder, UIApplicationDelegate {
     
     var helpViewController: HelpViewController?
-    var helpPresentedViewController: UIViewController?
     var helpButton: UIButton!
-                            
+
+    var webViewController: UIViewController?
+
     var window: UIWindow! = UIWindow(frame:CGRectMake(0, 0, UIScreen.mainScreen().bounds.height, UIScreen.mainScreen().bounds.width))
 
     func application(application: UIApplication!, didFinishLaunchingWithOptions launchOptions: NSDictionary!) -> Bool {
